init project
This commit adds 2 new template files CHANGELOG.CLIENT and
README.CLIENT. It also contains a README for this project.

These templates should be a start for every new project as they
already have the backbone and some explanation that projects share.

These templates are asciidoc files and can be generated as pdf or
html using the commands in README.CLIENT.

Issue: #83162, #83163
Change-Id: If3f79c0ebd7bc7e1c0e5a27ddc2f68b23302eef9