add docker build procedure
Add a procedure to build a project using docker. When starting a project with a
client, we need to select the architecture it has to work on. As developers have
heterogeneous environments, a docker build procedure will enhance the project
quality.

build_procedure contains the procedure to build with docker.
Dockerfile contains information about client's architecture.

Issue: #83165
Change-Id: Id21e5b3b5360a9f8c323a2e1b5a56ad2d673ea00