add script to build a project into docker
Add a script to build any project in docker using the client configuration.
As part as assuring better quality in a project development, we need a script
to automatically build into the client configuration. You just need a
Dockerfile with the client configuration. Default build command can be set in
.sflproject as well as release files to put in an archive.
This command can be overridden if you have your own build command as parameter.

Issue: #83331
Change-Id: Ib4561c347ee3aca0a7672f363dbeccaf302931c9