cqfd: redirect command stderr to stdout
When docker run without the -i -t mode, the docker run command
sometimes fails with the error "Unrecognized input header".

This error occurs when the buffer log is filled up. The first byte
is used to select where the logs should be written (stderr, stdout)
and seemed to be overwritten by the logs, triggering the error.

Looking at https://bugzilla.redhat.com/show_bug.cgi?id=1289476 a
possible workaround is to redirect stderr to stdout.

Change-Id: I178efa4b225b21684eb3dcdddc3d4d88ebae6cfa