cqfd: preserve environment in user shell
The initial code would not propagate the environment defined in the
Dockerfile into the user's shell, which would not be great in some
cases (e.g. when defining JAVA_HOME in Java projects).

This modifies the 'su' call to preserve the environment. Note that -p is
mutually exclusive with -l (aka su -).

Change-Id: I13aa77fa80b210b7c41d00a2293d3e9ce5a91561