cqfd: have Dockerfile path in config file
Dockerfile path is hard coded as ".cqfd/docker/Dockerfile". Create a
field in build section in .cqfdrc to use a Dockerfile of another distro.
To use a new Dockerfile, add a 'distro' key in the build section with
its name in it. The distro name have to match the repository for this
distro in .cqfd.
Example: a centos distro would have a Dockerfile in .cqfd/centos/ and
a distro='centos' build key
A new docker image will be generated only when the 'cqfd init' command
is used.