cqfd: print an error if ini-file is invalid
An error happens when cqfd was given an empty ini-file or when the ini-file
has an empty section.
cqfd/cqfd: eval: ligne 84: syntax error near unexpected token `}'

The line below causes the interpretor error:
	ini[${#ini[*]} + 1]='}'

It appends a brace which breaks to conversion from ini to shell.

In the case of an empty file, the conversion appends a useless closing
brace `}'; without any opening brace before `{'.

In the case of an empty section, the conversion appends an empty
function ``section() { }", which is not shell friendly. A null command
`:' must be placed between `{' and `}' to declare an empty function.