cqfd: check .cqfdrc existence at parse
Currently, the .cqfdrc file existence check is peformed at two places
into the script.

Thus, the test may be checked twice if the option -f is given through
the command-line.

This patch simplify the script by moving the test existence into the
parser; the only place where the file is used.