cqfd: add support for release filename templating
In the previous versions of cqfd, the name of the generated release file
was limited to include a Jenkins job name and build id, at most. This
was quite limiting in setups where users wanted the CI to generate more
elaborate archive filenames.

Limited (but working) release filename templating is added by this
commit to overcome this limitation. The user is now able to specify an
archive parameter in the release section of the .cqfdrc file, allowing
to use arbitrary environment variable names, as well as a few templating
macros in the resulting filename. The behavior and macros are described
in the documentation.

Tests have been adapted accordingly.