cqfd: allow multiline for files
Currently, the files property is not ready for a multiline
value.

Example of a multiline files property:
files='README.html \
       README.doc'

The configuration above results in the following shell error:
cqfd: fatal: Cannot create release: missing \

The `\' is not escaped and cqfd tries to test existence of a
file named `\' which does not exists.

This patch evaluates an echo fo the $files variable and
stores the output to $release_files variable.

Signed-off-by: Gaël PORTAY <gael.portay@savoirfairelinux.com>