cqfd: allow init to use flavor option
Currently the `init' command build an image without the flavor layer
while the `run' command does.

Despite the fact the code is rendered more symetric, the patch also
fixes unexpected behaviors if the flavor overwrites parameters used by
the config_load function (org, name and distro).

Considering the case of a flavor specify the Dockerfile (see distro
field).

	$ cat .cqfdrc
	[project]
	org='cqfd'
	name='test-distro'
	distro='centos'

	[build]
	flavors='ubuntu'
	command='grep "^NAME=" /etc/*release'

	[ubuntu]
	distro='ubuntu'

	$ cqfd -b ubuntu init
	sha256:2db69660930c565afc75566ae17ce502a688f4ad307d6ebf6cb18d95e02a5d37

will build a docker image without taking into consideration that the
flavor as changed the distro field. In that example, the image built
is a centos ; but ubuntu was expected.

	$ cqfd -b ubuntu
	useradd: warning: the home directory already exists.
	Not copying any file from skel directory into it.
	/etc/os-release:NAME="CentOS Linux"

The patch fixes this behavior:

	$ cqfd -b ubuntu init
	sha256:e2eea9aad800dc0b6a73e46ebcdb7760a03f5573313c47dd8e55e08976915ce2

	$ ../cqfd -b ubuntu
	/bin/bash: warning: setlocale: LC_ALL: cannot change locale (en_US.utf8)
	bash: warning: setlocale: LC_ALL: cannot change locale (en_US.utf8)
	/etc/os-release:NAME="Ubuntu"