cqfd: parser: allow to parse an empty ini file
This patch makes the parser smarter. It allows the parser to not fail
when it evaluates an empty ini file.

Instead, cqfd fails when it is about to use data from the cqfdrc that
are missing.

This is an example when the cqfdrc file is empty (using /dev/null).

Before this patch, cqfd was complaining about an invalid ini-file:

	$ cqfd -f /dev/null
	cqfd: fatal: /dev/null: Invalid ini-file!

... which does not help that much about the failure.

Now, it complains about a missing section:

	$ cqfd -f /dev/null
	cqfd: fatal: /dev/null: Missing project section!

... which may help the user to know what is going wrong.