cqfd: support multiple release archive formats
The .tar.gz and .zip archive formats are now supported, just specify the
proper file extension in the 'archive' key.

Since zip does not support rewriting paths in the archive, a new config
key, tar_transform, is added to switch that behaviour on.

The default behaviour is changed to keep the original paths in the
archive, whatever the archive format is.

This addresses #35 and #36.