use /tmp for temporary launch script
As reported by @logicielsjpb in #67, starting a command on OSX would
fail to map the launcher in the container, as OSX uses /var/folders/
instead of /tmp/ by default, preventing the script from being mapped.

This change mandates /tmp/ for the launcher script, which should be
fine from a POSIX perspective.