cqfdrc: add user_extra_groups command to cqfdrc
In some cases, the cqfd command requires an additional group
in order to execute correctly inside the container, like the
docker group. This commit adds a user_extra_groups which,
when specified, will append host's group of the same name to
the groups of cqfd's container user.
You can either use the 'group:gid' format, or simply specify
the group name if it exists either in the host or inside the
docker image.