cqfd: add support for alternative and external cqfd files.
The -f option was introduced to use alternative configuration files but
there was no option to use an alternative cqfd directory.

In addition, the -C option is a common option to change the working
directory in many unix tools. It allows the use of external
configuration files and can be useful when using cqfd in scripts.

This commit adds support for both the -d option and the -C option.