cqfd: add detection for changes in Dockerfile
When cqfd create an image, it is named "cqfd_username_org_project".
If the Dockerfile change, `cqfd init` recreate an image with the same name.

But if the user switch to a git branch with another Dockerfile
or if concurrent jobs use the same project
`cqfd run` will use an image that is not meant to him.

The goal of this commit it to synchronize `cqfd init` and `cqfd run` to ensure that the correct image is used.

To achieve this, the hash of the Dockerfile is put at the end of the image name.
`cqfd run` will display a new error message if this docker image is not found, requiring running `cqfd init` again.

This commit also adds a test corresponding to this feature.