cqfd: detect Dockerfile changes
Synchronize cqfd init and cqfd run to guarantee the correct Docker
image is used. Previously, cqfd init would label an image with
"cqfd_username_org_project" and recreate it using the same name if the
Dockerfile changed, leading to potential issues when switching
branches or using different Dockerfile revisions concurrently.

This commit appends the Dockerfile's hash to the image name and
implements an error message in `cqfd run` if the desired image is not
found, requiring the user to run `cqfd init` again.

A test has also been added to validate this feature.

 [jo: cleanup commit doc]