README: Add a warning about unused docker image
Cqfd named the docker image each time `cqfd init` is called. The old
naming system kept the same name for a new image if the Dockerfile had
changed, which let an unnamed and unused image that can be purged.

The new naming system including a Dockerfile hash use a new name every
time and thus let unused images with a name.
This can result in many unused images in the system.
This commit describes this problem in the user doc.