README: add a warning about unused docker image
Update README to include a warning regarding potential unused Docker
images. Previously, Cqfd maintained the same image name every time
cqfd init was run, which resulted in unreferenced images that could be
purged if the Dockerfile changed.

The new naming system appends a hash to the image name, creating a
unique name every time the Dockerfile changes, which can lead to
numerous unused images in the system.

This commit outlines this issue in the user documentation.

 [jo: fixup doc]