Makefile: fix completionsdir prefix
Both targets install and uninstall copy/remove the bash completion
script under the prefix /usr, whatever the value of $(PREFIX).

The pkg-config allows to redefine the variables thanks to the option
--define-variable; adding the option --define-variable=prefix=/usr/local
redefines the variable prefix to /usr/local.

Additionally, the generated file bash-completion.pc is broken somehow
since the commit 81ba2c7e7dfbaefbafa1e8615727c9612e5fb314 (version 2.9).
The content of the variable completionsdir hardcodes the content of the
prefix given at configure time instead of referencing it:

	completionsdir=/usr/local/share/bash-completion/completions

Instead of:

	completionsdir=${prefix}/share/bash-completion/completions

Note: Later, the variable datadir has been introduced with the commit
8c5302581ad20954cfa164dcae26f1ed6277dd90 (version 2.10), once again,
that variable hardcodes the content of the prefix value given a
configure time instad of referencing it:

	datadir=/usr/local/share
	completionsdir=${datadir}/bash-completion/completions

Instead of:

	datadir=${prefix}/share
	completionsdir=${datadir}/bash-completion/completions

Therefore, adding the option --define-variable=prefix=$(PREFIX) does not
work since bash-completion 2.9.

The variable datadir should be redefined to fix the hardcoding of the
prefix by adding --define-variable=datadir=${prefix}/share.

This adds the redefinition of the two variables prefix and datadir to
install the completion script to the appropriate completions directory
(i.e. using the given $(PREFIX)).