cqfd: reword error messages
cqfd now use the hash of the dockerfile to tag the docker image.
Thus, the Dockerfile is needed for the init task but also during the run task.

This new need can be confusing because the error message generated is
not relevant.

This commit rework the error messages to be more explicit.