Generate dynamically the list of flavors
The flavors are listed by the property flavors in the section build.

This list have to be maintained after each addition of a new flavor. It
it is a bit ennoying and kind of cumbersome if adding lots of flavors.

The compgen bash builtin allows to list the functions in the script (-A
function) and apply a pattern to filter the results (-X filterpat).

This generates the list of flavors dynamically after the .cqfdrc in file
is parsed. This filters out the sections project and build, and assumes
all other sections are flavors.