cqfd: add CQFD_NO_SSH_CONFIG option
The ssh configuration of the host may not be compatible with the ssh
version within the container. This option allows disabling ssh
configuration forwarding.

Because we might not bind ~/.ssh, the $cqfd_user_home may not exist
when this option is provided. I added a mkdir command to create it.

I chose not to include the SSH_AUTH_SOCK configuration in this option
because ~/.sockets/ssh have different compatibility.