docker_build: add throbber
As the cqfd --init command is going, in the future, to be ran
automatically, work has been done to make it less instrusive.

Rather than showing all of the Docker log, it just shows a throbber
with the current step status, and shows the full log only if the
docker build command failed.