tests: strip down centos7 Dockerfile
CentOS is discontinued.

Its Docker images are still available but the mirrors of the package
repositories are dying one after the others.

As a consequence, installing or updating packages lasts like an eternity
before the process ends with succeed or fails.

This strips down the centos Dockerfile to the bare minimun as there is
no reason to test for update packages.