Use $1 and $* instead of "$@"
The function docker_run() takes a single argument as the command string,
but it uses "$@". Therefore, $1 is enough and it should be used instead
to avoid confusion.

The shell expression "$@" expands to an array string, but the "$@" is
assigned to the scalar (i.e. non-array) variable build_cmd_alt;
Therefore, $* should be used instead.

This uses $1 and $* instead of "$@" at some places.