cqfd: fix syntax
Most of the functions in the script do not use a space between the name
and the parenthesis, and opens the curving brace in the same line.

This fixes syntax by removing the spaces between the function name and
its parenthesis and opening the curging brace in the same line for the
sake of consistancy.