cqfd: remove $build_cmd_append and use "$*"
The command run uses two variables to parse the arguments given to the
command line:
 - $build_cmd_append: to store the additional argument to concatenate to
   the command= set in the file .cqfdrc
 - $build_cmd_alt: to store the alterative command to replace entirely
   the command= set in the file .cqfdrc

Both of them store the given arguments thanks to the special parameter
$*.

Since the parsing of the argument is halted via the statement break, the
parameters remain available via $*. Thus, it is not needed to store them
in additional variables.

The parsing needs a flag to know if the command= in the file .cqfdrc is
to be replaced by the content of $* (i.e. cqfd run alt cmd) or to be
concatenated with the content of $* (i.e. cqfd run -c add args).

This removes entirely the variable $build_cmd_append and reuses the
variable $build_cmd_alt as a flag to either replace $build_cmd by $* or
to append $* to $build_cmd.