cqfd: restore IFS after call to parse_ini_config_file()
The function parse_ini_config_file() modifies (i.e. sets) the variable
IFS[1] and does not restore it.

As a consequence, the function modifies the word splitting[2] for the
entire script after it is called. The "words" are split by lines and
"breaks" the default word splitting by whitespaces.

This makes the variable IFS local to the function so the previous value
is automatically "restored" once the function exits.

[1]: https://www.gnu.org/software/bash/manual/bash.html#index-IFS
[2]: https://www.gnu.org/software/bash/manual/html_node/Word-Splitting.html