cqfd: fix tabulation character in variable expansions
The variable expansions intend to remove the tabulations before and
after the sign =, as per the comment.

However, the patterns match several spaces instead of a tabulation
(probably because of a copy/paste at code import).

This fixes the mismatch variable expansion by replacing the sequence of
spaces (backslash included) by a the single quoted string composed of
the ANSI escaped tabulation character (i.e. $'\t').