cqfd: Add option to forward the user's .gitconfig
When running commands that use git, it is useful to have the user's
.gitconfig available in the container. Mostly to have the user's name
and email configured.

For instance, when running Yocto's `devtool modify`, it will produce
an error if the user's name and email are not configured.

I chose to enable this new option by default, but it can be disabled
with the `CQFD_NO_USER_GIT_CONFIG` environment variable. It is
documented in the README.