cqfd: use CQFD_SHELL also for the default login shell
When running `cqfd run -c` commands, the default login shell could be
used instead of the one defined for `cqfd shell`. This was confusing
and not configurable. The same shell will be always used, defined by
the `CQFD_SHELL` environment variable.

This notably overcomes a regression when using scripts that require
bashisms. In the previous release, the default shell was `/bin/bash`,
and is now `/bin/sh`. Users relying on bashisms still need to set
`CQFD_SHELL` to `/bin/bash` in their environment.

See issue #310.