bash-completion: improve compatibility with zsh
When running bash-completion's completion functions in zsh, the
`_init_completion` and `_get_first_arg` functions are not defined in
`bashcompinit`. Defining dummy functions allow using auto-completion
in zsh although its not as accurate.

Without _get_first_arg, the completion doesn't detail the options
specific to that command and lists all of them.