cqfd: run: add warning if multiple arguments
The command 'run' is ambiguous.

This adds a warning if the command is run with multiple arguments
telling to use either 'exec' or 'shell'.

The warning is only printed when there is an argument containing a
whitespace, which is the tricky case.

See 344969c0de765faba5a22c8cb470e4efe63b2047.