cqfd: stop setting exit-code
exit(1) exits with the latest return-code if none is specified.

Therefore, the argument $? is spurious in exit $?.

Also, forcing to exit success using exit 0 and set -e is unreached as
the script has exited if the previous command fails.

Note: exit 1 (or any positive value up to 127) is the only "valid"
situation as it force the script to exit failure if the script has
reached that instruction.

This stops setting exit-codes for the two first situations.