cqfd: launcher: create temporary laucher to /tmp/cqfd-XXXXXX
The temporary launcher script is created to /tmp/tmp.XXXXXX and it is
removed on exit.

Some programs create files or directories with the exact same template.

This explicits the temporary launcher belongs to cqfd by updating the
file template to /tmp/cqfd_launcher.XXXXXX. It is easier to find the
remaining launcher files not removed this way.