cqfd: reverse test logic in function debug()
This reverses the test logic in function debug() of the launcher script
so it matches the other debug() function in cqfd script.

The pipeline of commands stops returning error if CQFD_DEBUG is unset.

With this change, the pipeline of commands returns success if CQFD_DEBUG
is unset, or return the output of the echo command if CQFD_DEBUG is set.

It is preferable the debug() wrapper does not end in error if the debug
traces are disabled, as it may ends the script if the option errexit is
set (i.e. set -e).

See:

	gportay@archlinux ~ $ unset CQFD_DEBUG
	gportay@archlinux ~ $ test -n "$CQFD_DEBUG" && echo set
	gportay@archlinux ~ $ echo $?
	1
	gportay@archlinux ~ $ CQFD_DEBUG=1
	gportay@archlinux ~ $ test -n "$CQFD_DEBUG" && echo set
	set
	gportay@archlinux ~ $ echo $?
	0

	gportay@archlinux ~ $ unset CQFD_DEBUG
	gportay@archlinux ~ $ test -z "$CQFD_DEBUG" || echo set
	gportay@archlinux ~ $ echo $?
	0
	gportay@archlinux ~ $ CQFD_DEBUG=1
	gportay@archlinux ~ $ test -z "$CQFD_DEBUG" || echo set
	set
	gportay@archlinux ~ $ echo $?
	0